requirements.txt
fastapi
uvicorn
jinja2
aiofiles
requests
websockets
apscheduler
sqlalchemy
RPi.GPIO

